Fix client event handlers registered in common
